+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 🔨 Other Changes
 
-- JSON import/export, duplicate handling, overwrite confirm ([83f63ba](../../commit/83f63ba))
-- Bump version to 0.7.0 in package.json, tauri.conf.json, and Cargo files ([f16e9c6](../../commit/f16e9c6))
+- Bump version to 0.7.1 in package.json, tauri.conf.json, and Cargo files ([c0b6cb1](../../commit/c0b6cb1))
 
 
-📋 Full Changelog: [f16e9c6...83f63ba](../../compare/f16e9c6...83f63ba)
+📋 Full Changelog: [c0b6cb1...c0b6cb1](../../compare/c0b6cb1...c0b6cb1)